Feature: Separate hooks from app view
