+++ src/App.jsx
@@ -1,16 +1,17 @@
-import { useState } from 'react';
 import './App.css';
 import { Movies } from './components/movies';
-
+import { useSearchQuery } from './hooks';
 
 function App() {
-  const [searchQuery, setSearchQuery] = useState("");
+  // controlled-Way
+  const { searchQuery, setSearchQuery, error } = useSearchQuery({ Query: "" })
+  console.log('Render', searchQuery)
 
   const handleSubmit = (e) => {
     e.preventDefault()
+    // Uncontrolled-Way
     // const formData = new FormData(e.target)
     // const { searchQuery } = Object.fromEntries(formData)
-
     console.log(searchQuery)
   }
 
@@ -19,15 +20,25 @@ function App() {
     setSearchQuery(newValue)
   }
 
+
+
   return (
     <div className='content-app'>
 
       <header>
         <h1>Search your movies</h1>
         <form onSubmit={handleSubmit} className='form-content'>
-          <input onChange={hanldeChange} type="search" name="searchQuery" id="searchQuery" placeholder='start wars, hallo, etc...' />
+          <input style={{
+            border: '1px solid transparent',
+            borderColor: error ? 'red' : 'transparent'
+          }} onChange={hanldeChange} type="search" name="searchQuery" id="searchQuery" placeholder='start wars, hallo, etc...' />
           <button type="submit">Buscar</button>
         </form>
+        <p style={{
+          color: 'red',
+        }}>
+          {error}
+        </p>
       </header>
 
       <main>
