+++ src/hooks/index.js
@@ -1,0 +1,2 @@
+export * from './useGetMovies'
+export * from './useSearchQuery'
